pyproject.toml
[tool.poetry]
name = "langchain-practice"
version = "0.1.0"
description = ""
authors = ["Nasir Abbas <akhtarabbas.islamicreaction@gmail.com>"]
readme = "README.md"
packages = [{ include = "langchain_practice" }]  # Ensure a valid module exists

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
langchain = "^0.3.17"
langchain-google-genai = "^2.0.9"
python-dotenv = "^1.0.1"
langchain-huggingface = "^0.1.2"
pypdf = "^5.2.0"
langchain-community = "^0.3.16"
faiss-cpu = "^1.10.0"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


